% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font_feature.R
\name{font_feature}
\alias{font_feature}
\title{Define OpenType font feature settings}
\usage{
font_feature(...)
}
\arguments{
\item{...}{key-value pairs with the key being the 4-letter tag and the value
being the setting (usually \code{TRUE} to turn it on).}
}
\value{
A \code{font_feature} object
}
\description{
This function encapsulates the specification of OpenType font features. Some
specific features have named arguments, but all available features can be
set by using its specific 4-letter tag For a list of the 4-letter tags
available see e.g. the overview on
\href{https://en.wikipedia.org/wiki/List_of_typographic_features}{Wikipedia}.
}
