% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_font.R
\name{match_font}
\alias{match_font}
\title{Find a system font by name and style}
\usage{
match_font(family, italic = FALSE, bold = FALSE)
}
\value{
A path locating the font file
}
\description{
This function locates the font file best matching a name and optional style
(italic/bold).
}
