% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemfonts-package.R
\docType{package}
\name{systemfonts-package}
\alias{systemfonts}
\alias{systemfonts-package}
\title{systemfonts: System Native Font Finding}
\description{
Provides system native access to the font catalogue. As font handling varies between systems it is difficult to correctly locate installed fonts across different operating systems. The 'systemfonts' package provides bindings to the native libraries on Windows, macOS and Linux for finding font files that can then be used further by e.g. graphic devices. The main use is intended to be from compiled code but 'systemfonts' also provides access from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/systemfonts}
  \item \url{https://systemfonts.r-lib.org}
  \item Report bugs at \url{https://github.com/r-lib/systemfonts/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})
  \item Devon Govett (Author of font-manager)
}

Other contributors:
\itemize{
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
