% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_fonts.R
\name{web-fonts}
\alias{web-fonts}
\alias{get_from_google_fonts}
\alias{get_from_font_squirrel}
\title{Download and add web font}
\usage{
get_from_google_fonts(family, dir = "~/fonts")

get_from_font_squirrel(family, dir = "~/fonts")
}
\arguments{
\item{family}{The font family to download (case insensitive)}

\item{dir}{Where to download the font to. The default places it in your user
local font folder so that the font will be available automatically in new R
sessions. Set to \code{tempdir()} to only keep the font for the session.}
}
\value{
A logical invisibly indicating whether a font was found and
downloaded or not
}
\description{
In order to use a font in R it must first be made available locally. These
functions facilitate the download and registration of fonts from online
repositories.
}
